% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Multispell data for HDM example}
\format{
A data frame with 128 rows and 39 variables:
}
\source{
\url{https://sites.psu.edu/forrest/heterogeneous-diffusion/}
}
\usage{
exampleData
}
\description{
A dataset covers a subsut of the data for adoptions of domestic partner benefits by US companies.
}
\keyword{datasets}
