% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdiffr.R
\name{hdiffr}
\alias{hdiffr}
\title{R functions for heterogeneous diffusion model}
\usage{
hdiffr(
  data,
  xvars,
  vvars = NULL,
  wvars = NULL,
  zdvars = NULL,
  zgvars = NULL,
  zmvar = NULL,
  zmvar2 = NULL,
  idvar,
  multispell = 0,
  repeatevents = 0,
  repeatvar = 0,
  vintercept = 1,
  robust = 0,
  hrno = 1,
  dist = "exponential"
)
}
\arguments{
\item{data}{data with SS (startstate), ES (endstate), ET (endtime), ST (starttime) columns (required)}

\item{xvars}{propensity varlist}

\item{vvars}{susceptibility varlist}

\item{wvars}{infectiousness varlist}

\item{zdvars}{distance-based proximity varlist}

\item{zgvars}{group-based proximity varlist}

\item{zmvar}{data for matrix-based proximity}

\item{zmvar2}{second data for matrix-based proximity}

\item{idvar}{focal ID}

\item{multispell}{option for multiple time spells per observation}

\item{repeatevents}{to use with multispell to allow IDs to be at-risk of adoption repeatedly}

\item{repeatvar}{to use with multispell and reveatvars to add a coefficient testing influence of repeat events in infectiousness}

\item{vintercept}{to test simple contagion}

\item{robust}{cluster errors based on ID}

\item{hrno}{to display output using coefficients rather than hazard ratios}

\item{dist}{to specify distribution for survreg (default is exponential)}
}
\value{
survival model results
}
\description{
hdiffr performs an event history diffusion model that is modified in order to
account for heterogeneity in the diffusion process. In addition to modeling
chracteristics that influence the direct propensity of a focal actor to adopt,
hdiffr can also model characteristics of the focal actor that makes it more
susceptible to the prior adoptions of other actors, as well as characteristics
of prior-adopting actors that make them more or less influential on others, and
characteristics of the social structure among actors that influence the transmission
of influence between actors.
}
\examples{
library(hdiffr)
data(exampleData)
result <- hdiffr(data = data, xvars = c('lnsale', 'roa', 'vote'),
      vvars = 'activistgrp', wvars = 'lnsale', zgvars = 'sic',
      idvar = 'gvkey', multispell = 1, vintercept = 1, hrno = 1)
}
